Make the Kits docker dev env behave more similar to our other Docker dev envs.
The way this works is that the docker-compose.yml file sets up a
dev env. When/if we start using this docker container in production,
we'll setup a heroku.yml file that passes in production values
through the ENV. Note that to actually have this all work, we need
to load a dev database and dev uploads/plugins. I'll be committing a
future change to add scripts for that.

TESTING:
With a dev DB loaded and uploads/plugins pulled from staging, I built
this container using docker-compose up -d and saw that the kits site
worked at kitsweb:3005. Note that I'm still having trouble getting it to
work using just kitsweb (with no port). For some reason, the nginx stuff
is acting up.