A slight hack to allow configureing the SSO Authorizer plugin to use CAS in a development environment without SSL support.
This is a slight hack, b/c it's not upgrade proof with the plugin. When
we upgrade, we'll have to fix this. Otherwise, I think it makes sense to
not assuming it's https by default if a port other than 443 or no port
is set. A real fix should be in the plugin where you can choose or it
goes off the current protocol.