Make the Docker container more closely match prod/staging.
Use the same Python version and set a php.ini config value that
prod/staging has which was masking a bug in our current version of
the Authorizer plugin where it can automatically redirect to the SSO
server by default. Without output_buffering on, headers are written
after data is written causing an error but the buffering masks it.
Upgrading to a newer Authorizer plugin would get rid of the need for
that setting I believe (but other bugs could crop up...)