Add support for the attendance tracking database in the Dev env.
Most of this commit is pulling the attendance dev DB, creating and
populating it. However, there is one piece related to actually being
able to login using the dev SSO server as well.

Note: to get the SSO settings right, I had to bypass SSO login
using: http://kitsweb:3005/wp-login?external-wordpress and login with
the local Wordpress admin account. I also had to fix up some php.ini
stuff to make it work (in a previous commit) as well as update the
settings in the database for the Authorizer plugin which is the SSO implementation.

TESTING:
- stopped the container, maunally removed the DB volume (since it's
  name, a docker-compose down -v doesn't remove it. You have to
explicitly delet it), rebuilt the container, ran docker-compose/scripts/dbrefresh.sh
and then went to http://kitsweb:3005 to see that it redirected me to the
local SSO dev env and I was able to login as a random user in that
database.
- I could see that the attendance database was populated by
logging into the DB and running queries against braven_attendance.
- Note: the attendance API (for canvas too pull the data into the
  gradebook) and the ability to actually pull the list of
  Fellows and LCs which drives the attendance tracking feature isn't quite done
yet. That's up next.