Get attendance tracking working in the dev env.
I had to regenerate the Canvas access token in the dev env, change the
DB refresh scripts on the Admin server to use the new encrypted values,
and reconfigure the dev env to use that. Note that any other app that
calls into the Canvas API will need to use this new token in the dev env
(aka Join server). I also had to make some code changes to choose the
protocol, either http or https based on what the current app was
running.

TESTING:
- Regenerated the dev DB from the Admin server (stored on the S3 bucket)
  and did a local dbrefresh.
- Rebuilt the kits docker dev env and did a dbrefresh.
- Logged in as the admin and went to SJSU kits that had attendance
  tracking embedded (all kits after the first 2).
- Saw that the Fellows in the course were listed with their attendance
  data and the ability to set it.