Make attendance API choose http vs https based on what the server is running
This is in preparation for if we create a way to run this in the dev env
but right now we aren't. It's consistent with what I did in the
attendance.php and attenance_shared.php and we have to test all of that
as we push this to the master branch, so we may as well get everything
working the same way all in one go.