Fix pushing attenance from kits to canvas gradebook.
A cronjob runs on the Kits server that runs the attendance_api.php
script. However, I had changed the shared login to check for the
server's http vs https situation. Since the cronjob is just a command
line thingie, none of that was set. Now, we choose the protocol based on
the FORCE_SSL_ADMIN config value in wp-config.php. If we have SSL on,
that should be set. This change looks a bit bigger than it should b/c
the code was using a buggy way to load wp-config values (the regex
didn't account for no space or more than one space between the key/value
and it also only handled values between single quotes, so boolean valued
configs weren't picked up.) Now it just loads the file direclty as a PHP
script and directly references the defined consts.

TESTING:
- I can still login as an LC and see the Fellows in my cohort and the
  value I choose for present vs absent is written to the database and
read on load.
- I ran `php attendance_api.php 2>&1` and it successfully calls in the
  canvas server. However, I'm still struggling to actuall get the
gradebook to populate b/c the event date/times for each Learning Lab
(aka event) are not setup properly. Working with the team to get that
done in prod and getting this PR up so I can test that last tricky bit
in staging b/c my dev env is stupid slow.