Add troubleshooting info for 'Undefined index: lc_email' error.
This error happens when there are sections without corresponding My Cohort Info
in the Portal. The code seems to still generally work but in DEBUG mode
it outputs a list of the above errors. This is confusing in staging.
Add some troubleshooting console.log's with more info so that staff testers
can get a sense of whether there is an actual issue with My Cohort Info or not.
In a future commit, we should change the actual logic of the code to handle this correctly
and not show the errors (or only show actual errors with more detail on what is wrong).
This is a stopgap.