Make Kits attendance handle hyphens in Kit names.
When a Kit is loaded for the first time with a [take-attendance]
shortcode, a corresponding event is created in the attendance_events
table using the Kit name. However, special characters like ampersands
and hyphens are HTML/URL encoded to their numeric value when they are
sent to the attendance code that handles this. We need to decode them
into normal strings so that the names will match up with the displayed
names over in the Portal calendar.

One tricky thing here is that the Kits database uses the Latin-1 character
set (aka ASCII). So any characters that are UTF8 which are not convertable
to ASCII would show up as a questions mark. So we're hacking it up and translating
any non-ASCII characters we encounter in Kit names to characters that can be stored
as ASCII chars. For now, we only handle the various types of hyphens. Also, we
do have kit names with an ampersand, but those decode fine using the normal method.

TESTING:
- Deleted an attendance event and reloaded that Kit. Saw that the new event
  was created without the numeric values, like &#8211;
- Took attendance for that Kit and saw that the attendance was saved.
- Updated the name of the event in the Portal calendar to match and ran attendance_api.php.
- Saw that the time from the calendar was pulled and stored on the event.
- Saw that the attendance I took was sent to the gradebook in the Portal.
- Did the same steps but with renaming an existing event instead of creating a new one.
  This is what we'll run in the database to clean up the existing stuff.